+++ wicket/docs-wicket/src/main/java/nl/rotterdam/wicket/docs/ComponentsPage.html
@@ -29,7 +29,7 @@
     <wicket:container wicket:id="badgeListExamplesPanel"></wicket:container>
     <wicket:container wicket:id="bodyExamplesPanel"></wicket:container>
     <wicket:container wicket:id="buttonExamplesPanel"></wicket:container>
-    <wicket:container wicket:id="buttonGroupExamplesPanel"></wicket:container>
+    <wicket:container wicket:id="actionGroupExamplesPanel"></wicket:container>
     <wicket:container wicket:id="codeBlockExamplesPanel"></wicket:container>
     <wicket:container wicket:id="columnLayoutExamplesPanel"></wicket:container>
     <wicket:container wicket:id="checkboxExamplesPanel"></wicket:container>
