+++ wicket/docs-wicket/src/main/java/nl/rotterdam/wicket/docs/action_group/ActionGroupExamplesPanel.html
@@ -8,14 +8,14 @@
           <p>Groepeert één of meer gerelateerde acties en verzorgt de lay-out van de content.</p>
         </div>
 
-        <section data-testid="buttonGroup">
+        <section data-testid="actionGroup">
           <h2>Button Group</h2>
           <div>
-            <p>Voorbeeld Button Group met één button.</p>
+            <p>Voorbeeld Action Group met één button.</p>
           </div>
           <rods-story-canvas>
             <template>
-              <div wicket:id="buttonGroup">
+              <div wicket:id="actionGroup">
                 <button wicket:id="button">voorbeeld knop</button>
               </div>
             </template>
